# sdk/go/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
servers:
- url: /
paths:
...
  /scopes:
    get:
      description: Get all scopes.
      operationId: get_scopes_scopes_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/ScopeViewModel"
                title: Response Get Scopes Scopes Get
                type: array
          description: Successful Response
      summary: Get Scopes
      tags:
      - Scopes
    post:
      description: Create a new scope.
      operationId: create_scope_scopes_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScopeCreateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScopeViewModel"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Create Scope
      tags:
      - Scopes
  /scopes/{scope_id}:
    delete:
      description: Delete a scope.
      operationId: delete_scope_scopes__scope_id__delete
      parameters:
      - explode: false
        in: path
        name: scope_id
        required: true
        schema:
          title: Scope Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                title: Response Delete Scope Scopes  Scope Id  Delete
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Delete Scope
      tags:
      - Scopes
    get:
      description: Get a scope by ID.
      operationId: get_scope_scopes__scope_id__get
      parameters:
      - explode: false
        in: path
        name: scope_id
        required: true
        schema:
          title: Scope Id
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScopeViewModel"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Scope
      tags:
      - Scopes
    put:
      description: Update an existing scope.
      operationId: update_scope_scopes__scope_id__put
      parameters:
      - explode: false
        in: path
        name: scope_id
        required: true
        schema:
          title: Scope Id
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScopeUpdateRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScopeViewModel"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Update Scope
      tags:
      - Scopes
...
components:
  schemas:
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          input: ""
          ctx: "{}"
          type: type
        - msg: msg
          loc:
          - Location_inner
          - Location_inner
          input: ""
          ctx: "{}"
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          title: Detail
          type: array
      title: HTTPValidationError
...
    MultiAgentSystemViewModel:
      description: View model for Multi-Agent System.
      properties:
        id:
          format: uuid
          title: Id
          type: string
        name:
          title: Name
          type: string
        authorization_server_id:
          format: uuid
          nullable: true
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
      required:
      - authorization_server_id
      - created_at
      - id
      - name
      title: MultiAgentSystemViewModel
    ScopeCreateRequest:
      description: Request model for scope creation.
      example:
        mas_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
      properties:
        name:
          title: Name
          type: string
        mas_id:
          format: uuid
          title: Mas Id
          type: string
      required:
      - mas_id
      - name
      title: ScopeCreateRequest
    ScopeUpdateRequest:
      description: Request model for scope updates.
      example:
        name: name
      properties:
        name:
          title: Name
          type: string
      required:
      - name
      title: ScopeUpdateRequest
    ScopeViewModel:
      description: Full view model for Scope including relationships.
      example:
        mas_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        tools:
        - input_schema: input_schema
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          scopes:
          - name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          output_schema: output_schema
          app_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - input_schema: input_schema
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          scopes:
          - name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - name: name
            id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          output_schema: output_schema
          app_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        mas: ""
      properties:
        id:
          format: uuid
          title: Id
          type: string
        name:
          title: Name
          type: string
        mas_id:
          format: uuid
          nullable: true
          type: string
        mas:
          allOf:
          - $ref: "#/components/schemas/MultiAgentSystemViewModel"
          nullable: true
        tools:
          items:
            $ref: "#/components/schemas/ToolViewModel"
          title: Tools
          type: array
      required:
      - id
      - mas
      - mas_id
      - name
      - tools
      title: ScopeViewModel
    ScopeViewModelMinimal:
      description: Minimal view model for Scope with only essential fields (used in
        Tool.scopes).
      example:
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          title: Id
          type: string
        name:
          title: Name
          type: string
      required:
      - id
      - name
      title: ScopeViewModelMinimal
...
    ToolViewModel:
      description: View model for Tool including its scopes.
      example:
        input_schema: input_schema
        name: name
        description: description
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        scopes:
        - name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - name: name
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        output_schema: output_schema
        app_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          title: Id
          type: string
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        input_schema:
          title: Input Schema
          type: string
        output_schema:
          title: Output Schema
          type: string
        app_id:
          format: uuid
          nullable: true
          type: string
        scopes:
          items:
            $ref: "#/components/schemas/ScopeViewModelMinimal"
          title: Scopes
          type: array
      required:
      - app_id
      - description
      - id
      - input_schema
      - name
      - output_schema
      - scopes
      title: ToolViewModel
...
    ValidationError:
      example:
        msg: msg
        loc:
        - Location_inner
        - Location_inner
        input: ""
        ctx: "{}"
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/Location_inner"
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
        input:
          title: Input
        ctx:
          title: Context
          type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    Location_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http